- made lint clean.
git-svn-id: file:///svn/unbound/trunk@3050 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ daemon/daemon.h
@@ -76,7 +76,7 @@ struct daemon {
 	 * or just one element[0] shared by the worker threads. */
 	struct listen_port** ports;
 	/** size of ports array */
-	int num_ports;
+	size_t num_ports;
 	/** port number for remote that has ports opened. */
 	int rc_port;
 	/** listening ports for remote control */
